third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/forms/the-option-element/option-text-label-expected.txt

This is a testharness.js-based test.
FAIL Option with non-empty label. assert_equals: expected "text" but got "label"
PASS Option with empty label. 
Harness: the test ran to completion.
